New build system.

+++ exe/Jhu.Graywulf.RemoteService.Server/post-build.ps1
@@ -1,18 +1,8 @@
-﻿param(
-	[string]$SolutionDir,
-	[string]$SolutionName,
-	[string]$ProjectDir,
-	[string]$ProjectName,
-	[string]$OutDir,
-	[string]$TargetName
-)
-
-cp $ProjectDir$OutDir$TargetName.exe $SolutionDir$OutDir
+﻿cp $ProjectDir$OutDir$TargetName.exe $SolutionDir$OutDir
 cp $ProjectDir$OutDir$TargetName.pdb $SolutionDir$OutDir
 cp $ProjectDir$OutDir$TargetName.exe.config $SolutionDir$OutDir
 
-cp "$ProjectDir..\..\util\eseutil.exe" "$ProjectDir$OutDir"
-cp "$ProjectDir..\..\util\ese.dll" "$ProjectDir$OutDir"
-
-cp "$ProjectDir..\..\util\eseutil.exe" "$SolutionDir$OutDir"
-cp "$ProjectDir..\..\util\ese.dll" "$SolutionDir$OutDir"
+if ( -not (test-path "$SolutionDir$OutDir\eseutil.exe") ) {
+  cp "$ProjectDir..\..\util\eseutil.exe" "$SolutionDir$OutDir"
+  cp "$ProjectDir..\..\util\ese.dll" "$SolutionDir$OutDir"
+}
